created registration endpoint
